fixed variable mix up with mc_ip and url for mc website
added file check for mc_log function, make sure file exists. also added random number to check status command sent to server. Checks status of mc server by sending 'STATUS' string to server (isn't actual command), then checks logs whether it was logged, if so it's active, if not the command was probably sent to a shell instead of server, so it won't show up in the log. But the issue was sometimes it'll find 'STATUS' in log but from a older session, even if server isn't actually active. So I added a random number to each status command sent, so it doesn't mistakenly find older commands in log and say the server is active when it's not. prob could be a better way to do it, lmk.