changed some argument, and added startbot, startserver, startboth
startbot just starts the bot, and startserver just starts the mc server and startboth starts.....both, and it uses either tmux or ssubprocess depending on booleans set in server_functions.py. if not using tmux and want to start bot, it'll start the bot in the current shell, and will have to use discord command to start mc server.