changed log time format. status checker now works with PaperMC
Regular minecraft server, if inputted was not a command it would log it anyways, with PaperMC, it has to be a valid command to log it, even if command failed or arguments are wrong. Before I was just sending STATUS with a random number, so if the server was active it'll log it, and i'll check the log, but now that doesn't work with everything. So I use debug then statuschecker plus the random number, so PaperMC logs it.