Trying to fix up and touch up RCON functionality.
Changed it so each time the bot needs to send a command via RCON it'll open the connection and close it right after, which will happen multiple times due to the whole send random number functionality, I would like to streamline it in the future, but for now it seems to work. 

If minecraft server is found to be offline, new server_active boolean will be updated, the bot won't try to send any commands to the server if boolean is false. To have the bot check on the status of the server use ?stats or new ?check command, which will update the server_active boolean so the other commands will work now. Probably not the best method of doing this....