Added ?check command to command_info.csv
idk how to word it since it is confusing how ?stats and ?check work when server is offline. If you try any command and the server is offline, it'll update the server_active boolean which will make it so all the other commands won't go ahead and try to command the minecraft server, Most commands won't actually check server status, ?check and ?stats will however actually check active status and if it's online it'll update the server_active boolean so other commands can work.