Changed lprint bot log text, and updated how autosave loop works
Autosave loop will now just run in the background at bot startup, you can't stop or start the loop. Instead at each loop it'll use server_command() to both check and send command. This integrates with server_active boolean, so if suddenly at one loop and the server is now offline, it'll update server_active bool and the rest of the loops it won't even send the command, which is just how server_command works now. And use ?stats or ?check to update the bot on server active state.