New restore panels, and able to delete components
Panels to restore/delete world or server backups.
Everytime worldrestorepanel or serverrestorepanel is called, selection variables are cleared and _delete_current_components is called. This is to avoid conflicts. e.g. before, you could have multiple components in history with different values (listings), but the button all does the same, so you could accidentally delete a wrong backup or something. 
Other commands don't really need this because they don't manipulate game files so it's less risky and doesn't matter as much if run in to conflict.