Fixed buffer time, wasn't implemented correctly
I only added the wait time to the command that sends the random number, but after that it will send the command and not wait.
So i added it, sadly now the commands may take longer, since it has to send a command with random number to check server status, then send the actual command user wanted to send.