Fixed unable to get players for 1.12 and lower
Server versions 1.12 and below outputs two lines for /list command instead of 1. First line is the 'There are x players' message, and the second line is the list of player names. Newer servers (1.13+) has it all in one line.

New server_version user config. The bot will try to detect the server version itself, but if it doesn't work or you want to manually set the server version yourself you can now.
NOTE: This will set the bot compatibly for server version for all servers you may have. So if you switch servers (with ?selectserver) I would not recommend setting this, Defaults to None.

I've only tested this for PaperMC (it should work for vanilla tho).

Some other fixes relating to server_status().